D-Bug-Django/requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
cffi==1.15.1
charset-normalizer==2.0.12
cloudinary==1.30.0
commonmark==0.9.1
cryptography==38.0.3
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.5
django-oauth-toolkit==2.2.0
djangorestframework==3.13.1
httpie==3.2.1
idna==3.3
jwcrypto==1.4.2
multidict==6.0.2
oauthlib==3.2.2
psycopg2==2.9.3
pycparser==2.21
Pygments==2.12.0
PySocks==1.7.1
python-decouple==3.6
pytz==2022.1
requests==2.28.0
requests-toolbelt==0.9.1
rich==12.4.4
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.2.0
urllib3==1.26.9
wrapt==1.14.1
